+++ CHANGES.md
@@ -28,6 +28,14 @@ development source code and as such may not be routinely kept up to date.
   Batch job.
   ([#460](https://github.com/nextstrain/cli/pull/460))
 
+* When Nextstrain CLI detects that a new version of itself is available, it
+  will now output the instructions on how to upgrade to that new version after
+  (instead of before) the normal output from the running command.  This makes
+  it easier to find and follow the instructions when output is long.  The
+  notices of the new version will be output both before and after normal
+  output.
+  ([#475](https://github.com/nextstrain/cli/issues/475))
+
 ## Bug fixes
 
 * `nextstrain setup <pathogen>@<version>` and `nextstrain update <pathogen>@<version>`